+++ src/entities/City.java
@@ -2,13 +2,13 @@ package entities;
 
 // Não utiliza o hibernate pois utilizará o arquivo 'states-cities.json' para consultas
 public class City {
-    private Long id;
+    private int id;
     private Long stateId;
     private String name;
 
     public City() {}
 
-    public City(String name, Long id, Long stateId) {
+    public City(String name, int id, Long stateId) {
         this.name = name;
         this.id = id;
         this.stateId = stateId;
@@ -22,11 +22,11 @@ public class City {
         this.name = name;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
